feat(webos): move date-time to navbar, remove from dock
Inline time/date display into the navbar right side as plain text.
Removes the standalone DateTimeDisplay pill component and its usage
in the taskbar, keeping the dock clean.
